Refactor CLI tooling into src/cli modules
